Carrito Rework

+++ app/src/main/java/com/example/myapplication/ComandaManager.kt
@@ -36,7 +36,7 @@ object ComandaManager {
     data class ContingutAdd(
         val id: Int,
         val nom: String,
-        val preuTotal: Int,
+        val preuTotal: Float,
         val quantitat: Int
     )
 }
